test: get version after EGLContext is ready
